compilers: put elbrus in mixins
