+++ mesonbuild/compilers/c.py
@@ -25,12 +25,12 @@ from .mixins.visualstudio import VisualStudioLikeCompiler
 from .mixins.gnu import GnuCompiler
 from .mixins.intel import IntelGnuLikeCompiler, IntelVisualStudioLikeCompiler
 from .mixins.clang import ClangCompiler
+from .mixins.elbrus import ElbrusCompiler
 from .compilers import (
     gnu_winlibs,
     msvc_winlibs,
     Compiler,
     CompilerType,
-    ElbrusCompiler,
     PGICompiler,
 )
